+++ doc/version-history.rst
@@ -10,6 +10,12 @@ Version History
 .. No new work should be required in order to complete this section.
 .. Below is an example of a version history format.
 
+v0.17.0
+-------
+* Parameterized the OCSP 2||3 index, determined by test environment.
+* Updated love_stress_test to define the LOVE 'location' URL based on the test environment.
+* Added LOVE kubernetes instance testing.
+
 v0.16.0
 -------
 * Added Watcher to the ObsSys State transition tests.